rename useGetTags to useGetDanbooruTags
